+++ libgadget/tests/test_slotsmanager.c
@@ -23,7 +23,7 @@ setup_particles(void ** state)
     PartManager->MaxPart = 1024;
     PartManager->NumPart = 128 * 6;
 
-    int newSlots[6] = {128, 128, 128, 128, 128, 128};
+    int64_t newSlots[6] = {128, 128, 128, 128, 128, 128};
 
     PartManager->Base = (struct particle_data *) mymalloc("P", PartManager->MaxPart* sizeof(struct particle_data));
     memset(PartManager->Base, 0, sizeof(struct particle_data) * PartManager->MaxPart);
@@ -41,7 +41,7 @@ setup_particles(void ** state)
 
     slots_setup_topology(PartManager, newSlots, SlotsManager);
 
-    int i;
+    int64_t i;
     #pragma omp parallel for
     for(i = 0; i < PartManager->NumPart; i ++) {
         P[i].ID = i;
@@ -109,8 +109,8 @@ test_slots_reserve(void **state)
      * moving those numbers to All.* we shall rework the code here. */
     setup_particles(state);
 
-    int newSlots[6] = {128, 128, 128, 128, 128, 128};
-    int oldSize[6];
+    int64_t newSlots[6] = {128, 128, 128, 128, 128, 128};
+    int64_t oldSize[6];
     int ptype;
     for(ptype = 0; ptype < 6; ptype++) {
         oldSize[ptype] = SlotsManager->info[ptype].maxsize;
